- Move API key retrieval method to API Key service
