+++ src/services/api_key_service.js
@@ -1,4 +1,8 @@
 import baseService from './base_service';
+import Constants from '../constants';
+import FilteringError from '../errors/general/filtering_error';
+
+const util = require('util');
 
 export default (easypostClient) =>
   /**
@@ -15,5 +19,35 @@ export default (easypostClient) =>
       const url = 'api_keys';
 
       return this._all(url, params);
+    }
+
+    /**
+     * Retrieve API Keys for a specified {@link User user}.
+     * See {@link https://www.easypost.com/docs/api/node#retrieve-an-api-key EasyPost API Documentation} for more information.
+     * @param {string} id - The ID of the user to retrieve keys for.
+     * @returns {Array} - List of associated API Keys.
+     * @throws {FilteringError} If user or API Keys are not found.
+     */
+    static async retrieveApiKeysForUser(id) {
+      const url = `api_keys`;
+
+      try {
+        const response = await easypostClient._get(url);
+        const user = this._convertToEasyPostObject(response.body);
+
+        if (user.id == id) {
+          return user.keys;
+        }
+
+        user.children.forEach((child) => {
+          if (child.id == id) {
+            return child.keys;
+          }
+        });
+      } catch (e) {
+        return Promise.reject(e);
+      }
+
+      throw new FilteringError({ message: util.format(Constants.NO_OBJECT_FOUND, 'child') });
     }
   };
